clean up cli
